+++ src/psd.h
@@ -3,7 +3,7 @@
 
 #include <complex.h>
 
-void calculatePSD (double complex *in, int num_samples,
+void calculatePSD (float complex *in, int num_samples,
                    double sampling_rate, double *psd);
 
 #endif
